Initial Commit
This initial commit turns the repository into a hugo module
which can be added to a Platen site to set it up with Netlify
CMS; it only handles adding and editing blog posts.

Future commits will be needed to extend the configuration to
handle basic pages, configuration, and game content.